+++ src/extensions/application-customizer.manifest.json
