+++ src/extensions/application-customizer.module.scss
