--- src/extensions/components/home-button.int.ts
@@ -1,5 +1,0 @@
-import { IPortalContext } from "../../common/portal-context";
-
-export interface IHomeButtonProps {
-    
-}
