+++ src/extensions/components/home-button.interfaces.ts
@@ -1,0 +1,5 @@
+import { IPortalContext } from "../../common/portal-context.interfaces";
+
+export interface IHomeButtonProps {
+    
+}
