--- src/extensions/components/portal-container.int.ts
@@ -1,10 +1,0 @@
-import { WebPartContext } from "@microsoft/sp-webpart-base";
-import { IPortalContext } from "../../common/portal-context";
-
-export interface IPortalProps {
-    portalContext: IPortalContext;
-}
-
-export interface IPortalState {
-
-}
