+++ src/extensions/components/portal-container.interfaces.ts
@@ -1,0 +1,9 @@
+import { IPortalContext } from "../../common/portal-context.interfaces";
+
+export interface IPortalProps {
+    portalContext: IPortalContext;
+}
+
+export interface IPortalState {
+
+}
